requirements.txt
pandas==1.3.3
plotly==5.5.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
requests==2.26.0
python-binance==1.0.15
gunicorn==20.1.0
plotly_express==0.4.0
dash_bootstrap_components==0.12.0